+++ source/index.html
@@ -350,8 +350,8 @@
               <source srcset="img/result1-mob.png 1x, img/result1-mob@2x.png 2x" media="(max-width: 768px) and (orientation: portrait)">
               <img src="img/result1.png" srcset="img/result1@2x.png 2x" alt="">
             </picture> -->
-            <canvas class="result__canvas" id="sea-calf"></canvas>
-            <canvas class="result__canvas" id="airplane"></canvas>
+            <canvas class="result__canvas" id="sea-calf">Обновите браузер</canvas>
+            <!-- <canvas class="result__canvas" id="airplane"></canvas> -->
           </div>
           <h2 class="result__title">
             <span class="visually-hidden">ПОбеда!</span>
